Renamed some folders and files
