+++ pyBumpHunter/bumphunter_1dim.py
