+++ setup.cfg
@@ -33,7 +33,7 @@ install_requires =
 python_requires = >=3.6
 
 [options.packages.find]
-exclude = testing
+exclude = test
 
 [options.extras_require]
 example =
