+++ test/test_BumpHunter.py
